Add Bitbucket Server Urls validation by pattern match (#761)
When we validate a Bitbucket Server url but neither oauth nor Personal Access Token is set up, we do a test Bitbucket Server Api request and if we get response with status 200 or 401 unauthorized, it means that the url is a Bitbucket Server url. The problem is that other unsupported git providers may also return such response on the test request. To make the check more specific, additionally test the url to match the Bitbucket Server url pattern.