Support GitHub repository devfile url (#793)
Add an ability to start a workspace from the repository url of a devfile e.g. https://github.com/vinokurig/test/blob/master/devfile.yaml This will be an experimental feature and will be applied only for GitHub.

Since GitHub url defines branch as a segment of the file path e.g. https://github.com/user/repository/blob/branch/file, we can not distinguish a branch with / from the path. For instance we can have such path: /branch/with/slash/file in the url, so we can not parse it as <branch>/<path to file> as both branch and path may contain /, see.

In this implementation when parsing a file path from a url, we give the priority to the file path, so when parsing /branch/with/slash/file, we assume that branch => branch/with/slash and file path => file

With this solution we can generate RAW devfile url from a repository url like: https://github.com/user/repository/blob/branch/path/to/devfile.yaml, but we pass branch/path/to to the factory parameters, and as a result the project clone will fail and the default branch will be used. However, we do not break the existing functionality that allows to start a workspace with branch that has / in its name.

---------

Co-authored-by: Aleksandr Shmaraiev <oshmarai@redhat.com>